requirements.txt
Flask==0.10.1
Flask-Cache==0.13.1
mysql-connector-python==2.0.4
